Now shows string value name above object in AR

+++ app/src/main/java/io/synople/scanmoney/model/Money.kt
@@ -1,5 +1,5 @@
 package io.synople.scanmoney.model
 
-data class Money(val x: Double,
-                 val y: Double,
+data class Money(val x: Float,
+                 val y: Float,
                  val value: String)
